#832: fixed some unit tests
