+++ tests/data_subscriber/conftest.py
@@ -1,0 +1,1 @@
+../unit/conftest.py
